requirements.txt
annotated-types==0.7.0
anyio==4.5.2
asttokens==3.0.0
backcall==0.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
contourpy==1.1.1
cycler==0.12.1
decorator==5.1.1
dlib==19.24.6
exceptiongroup==1.2.2
executing==2.1.0
fastapi==0.115.6
filelock==3.16.1
fonttools==4.55.1
fsspec==2024.10.0
gitdb==4.0.11
GitPython==3.1.43
h11==0.14.0
httptools==0.6.4
idna==3.10
importlib-resources==6.4.5
ipython==8.12.3
jedi==0.19.2
jinja2==3.1.4
kiwisolver==1.4.7
MarkupSafe==2.1.5
matplotlib==3.7.5
matplotlib-inline==0.1.7
mpmath==1.3.0
networkx==3.1
numpy==1.24.4
opencv-python==4.10.0.84
packaging==24.2
pandas==2.0.3
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.4.0
prompt-toolkit==3.0.48
psutil==6.1.0
ptyprocess==0.7.0
pure-eval==0.2.3
py-cpuinfo==9.0.0
pydantic==2.10.3
pydantic-core==2.27.1
pygments==2.18.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
scipy==1.10.1
seaborn==0.13.2
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
stack-data==0.6.3
starlette==0.41.3
sympy==1.13.3
torch==2.4.1
torchvision==0.19.1
tqdm==4.67.1
traitlets==5.14.3
typing-extensions==4.12.2
tzdata==2024.2
ultralytics==8.3.40
ultralytics-thop==2.0.12
urllib3==2.2.3
uvicorn==0.32.1
uvloop==0.21.0
watchfiles==0.24.0
wcwidth==0.2.13
websockets==13.1
zipp==3.20.2
